DRY trace event names
